+++ mercury/mercury-library/agent-cli-didcommx/src/main/scala/io/iohk/atala/mercury/AgentCli.scala
@@ -3,20 +3,20 @@ package io.iohk.atala.mercury
 import scala.jdk.CollectionConverters.*
 import zio.*
 
+import zio._
+import zio.http._
+import zio.http.model._
+import zio.http.service._
 import java.io.IOException
-import io.iohk.atala.resolvers.PeerDidMediatorSecretResolver
-import zhttp.service.ChannelFactory
-import zhttp.service.EventLoopGroup
 import org.didcommx.didcomm.DIDComm
-import io.iohk.atala.resolvers.UniversalDidResolver
-import zhttp.service.*
-import zhttp.http.*
 import io.iohk.atala.QRcode
-import io.iohk.atala.mercury.model.{*, given}
-import io.iohk.atala.mercury.model.error.*
-import io.iohk.atala.mercury.protocol.outofbandlogin.*
-import io.iohk.atala.mercury.protocol.issuecredential.*
-import io.iohk.atala.mercury.protocol.presentproof.*
+import io.iohk.atala.mercury.model.{_, given}
+import io.iohk.atala.mercury.model.error._
+import io.iohk.atala.mercury.protocol.outofbandlogin._
+import io.iohk.atala.mercury.protocol.issuecredential._
+import io.iohk.atala.mercury.protocol.presentproof._
+import io.iohk.atala.resolvers.PeerDidMediatorSecretResolver
+import io.iohk.atala.resolvers.UniversalDidResolver
 import io.iohk.atala.mercury.protocol.connection.*
 import io.iohk.atala.mercury.protocol.invitation.v2.Invitation
 
@@ -57,7 +57,7 @@ object AgentCli extends ZIOAppDefault {
       |DID Comm V2 Agent - CLI tool for debugging - Build by Atala (IOHK)
       |""".stripMargin)
 
-  val env = ChannelFactory.auto ++ EventLoopGroup.auto()
+  // val env = zio.http.Client.default ++ zio.Scope.default
 
   def agentLayer(peer: PeerDID): ZLayer[Any, Nothing, AgentServiceAny] = ZLayer.succeed(
     io.iohk.atala.mercury.AgentServiceAny(
@@ -75,7 +75,8 @@ object AgentCli extends ZIOAppDefault {
       }
       _ <- CoordinateMediationPrograms
         .senderMediationRequestProgram(mediatorURL = url)
-        .provideSomeLayer(env)
+      // .provideSomeLayer(zio.http.Client.default)
+      // .provideSomeLayer(zio.Scope.default)
     } yield ()
   }
 
@@ -147,7 +148,8 @@ object AgentCli extends ZIOAppDefault {
           content = Body.fromChunk(Chunk.fromArray(jsonString.getBytes)),
           // ssl = ClientSSLOptions.DefaultSSL,
         )
-        .provideSomeLayer(env)
+        .provideSomeLayer(zio.http.Client.default)
+        .provideSomeLayer(zio.Scope.default)
       data <- res.body.asString
       _ <- Console.printLine(data)
     } yield ()
@@ -229,7 +231,6 @@ object AgentCli extends ZIOAppDefault {
 
     import io.iohk.atala.mercury.protocol.invitation.OutOfBand
     import io.circe._, io.circe.parser._
-    import io.iohk.atala.mercury.protocol.invitation.InvitationCodec._
     for {
       didCommService <- ZIO.service[DidComm]
       _ <- Console.printLine("Read OutOfBand Invitation")
@@ -285,7 +286,8 @@ object AgentCli extends ZIOAppDefault {
           content = Body.fromChunk(Chunk.fromArray(jsonString.getBytes)),
           // ssl = ClientSSLOptions.DefaultSSL,
         )
-        .provideSomeLayer(env)
+        .provideSomeLayer(zio.http.Client.default)
+        .provideSomeLayer(zio.Scope.default)
         .catchNonFatalOrDie { ex => ZIO.fail(SendMessage(ex)) }
       data <- res.body.asString
         .catchNonFatalOrDie { ex => ZIO.fail(ParseResponse(ex)) }
@@ -298,7 +300,8 @@ object AgentCli extends ZIOAppDefault {
     Http
       .collectZIO[Request] {
         case req @ Method.POST -> !!
-            if req.headersAsList.exists(h => h._1.equalsIgnoreCase(header._1) && h._2.equalsIgnoreCase(header._2)) =>
+            if req.headersAsList
+              .exists(h => h._1.toString.equalsIgnoreCase(header._1) && h._2.toString.equalsIgnoreCase(header._2)) =>
           req.body.asString
             .catchNonFatalOrDie(ex => ZIO.fail(ParseResponse(ex)))
             .flatMap { data =>
@@ -334,7 +337,7 @@ object AgentCli extends ZIOAppDefault {
       case ""  => ZIO.succeed(defualtPort)
       case str => ZIO.succeed(str.toIntOption.getOrElse(defualtPort))
     }
-    _ <- Server.start(port, webServer(port)).debug.fork
+    _ <- Server.serve(webServer(port)).debug.fork
     _ <- Console.printLine("Endpoint Started")
   } yield ()
 
@@ -362,14 +365,14 @@ object AgentCli extends ZIOAppDefault {
     } yield (peer)
 
     didCommLayer = agentLayer(agentDID)
-    layers: ZLayer[Any, Nothing, AgentServiceAny & HttpClient] = didCommLayer ++ HttpClientZhttp.layer
+    layers: ZLayer[Any, Nothing, AgentServiceAny & HttpClient] = didCommLayer ++ ZioHttpClient.layer
 
     _ <- options(
       Seq(
         "none" -> ZIO.unit,
         "Show DID" -> Console.printLine(agentDID),
         "Get DID Document" -> Console.printLine("DID Document:") *> Console.printLine(agentDID.getDIDDocument),
-        "Start WebServer endpoint" -> startEndpoint.provide(didCommLayer),
+        "Start WebServer endpoint" -> startEndpoint.provide(zio.http.Server.default, didCommLayer),
         "Ask for Mediation Coordinate" -> askForMediation.provide(layers),
         "Generate login invitation" -> generateLoginInvitation.provide(didCommLayer),
         "Login with DID" -> loginInvitation.provide(didCommLayer),