+++ mercury/mercury-library/protocol-invitation/src/main/scala/io/iohk/atala/mercury/protocol/invitation/v2/Invitation.scala
@@ -1,10 +1,10 @@
 package io.iohk.atala.mercury.protocol.invitation.v2
-import cats.implicits._
-import io.circe.syntax._
-import io.circe.{Encoder, Json}
-import io.iohk.atala.mercury.model._
+import cats.implicits.*
+import io.circe.syntax.*
+import io.circe.{Encoder, Decoder, Json}
+import io.iohk.atala.mercury.model.*
 import AttachmentDescriptor.attachmentDescriptorEncoderV2
-import io.iohk.atala.mercury.protocol.invitation.InvitationCodec._
+import io.circe.generic.semiauto._
 
 /** Out-Of-Band invitation
   * @see
@@ -14,7 +14,7 @@ final case class Invitation(
     id: String = java.util.UUID.randomUUID.toString(),
     `type`: PIURI = Invitation.`type`,
     from: DidId,
-    body: Body,
+    body: Invitation.Body,
     attachments: Option[Seq[AttachmentDescriptor]] = None
 ) {
   assert(`type` == "https://didcomm.org/out-of-band/2.0/invitation")
@@ -23,11 +23,21 @@ final case class Invitation(
 }
 
 object Invitation {
-  def `type`: PIURI = "https://didcomm.org/out-of-band/2.0/invitation"
-}
 
-case class Body(
-    goal_code: String,
-    goal: String,
-    accept: Seq[String]
-)
+  final case class Body(
+      goal_code: String,
+      goal: String,
+      accept: Seq[String]
+  )
+
+  object Body {
+    given Encoder[Body] = deriveEncoder[Body]
+
+    given Decoder[Body] = deriveDecoder[Body]
+  }
+
+  def `type`: PIURI = "https://didcomm.org/out-of-band/2.0/invitation"
+  given Encoder[Invitation] = deriveEncoder[Invitation]
+  given Decoder[Invitation] = deriveDecoder[Invitation]
+
+}