+++ pollux/lib/core/src/main/scala/io/iohk/atala/pollux/core/model/error/CreateCredentialPayloadFromRecordError.scala
@@ -1,0 +1,8 @@
+package io.iohk.atala.pollux.core.model.error
+
+sealed trait CreateCredentialPayloadFromRecordError
+
+object CreateCredentialPayloadFromRecordError {
+  final case class RepositoryError(cause: Throwable) extends CreateCredentialPayloadFromRecordError
+  final case class CouldNotExtractClaimsError(cause: Throwable) extends CreateCredentialPayloadFromRecordError
+}