+++ pollux/lib/sql-doobie/src/main/resources/sql/pollux/V1__init_tables.sql
@@ -15,12 +15,15 @@
 
 CREATE TABLE public.issue_credential_records(
   "id" VARCHAR(36) NOT NULL PRIMARY KEY,
+  "created_at" BIGINT NOT NULL,
+  "updated_at" BIGINT,
   "thid" VARCHAR(36) NOT NULL,
   "schema_id" VARCHAR(36),
-  "subject_id" TEXT NOT NULL,
   "role"  VARCHAR(50) NOT NULL,
+  "subject_id" TEXT NOT NULL,
   "validity_period" INTEGER,
-  "claims" TEXT NOT NULL,
+  "automatic_issuance" BOOLEAN,
+  "await_confirmation" BOOLEAN,
   "protocol_state" VARCHAR(50) NOT NULL,
   "publication_state" VARCHAR(50),
   "offer_credential_data" TEXT,